\name{maturationData}
\docType{data}
\alias{maturationData}
\title{Maturation Data}
\description{
A five-dimensional array of dimensions time (in minutes) x data columns x 
sample x view x tFT (identified by the RFP since sfGFP is present for all) 
containing fluorescence intensity readouts for the fluorophore maturation 
curves.
}
\usage{maturationData}
\keyword{datasets}
