\name{runShinyApps}
\alias{runShinyApps}
\alias{runChooseFP2App}
\alias{runTimerModellingApp}
\title{
Run R-shiny applications
}
\description{
Wrapper functions that run shiny apps located in extdata subdirectory of R 
package.
}
\usage{
runChooseFP2App()
runTimerModellingApp()
}
\author{
Joseph D. Barry
}
\examples{
\donttest{
runChooseFP2App()
}
\donttest{
runTimerModellingApp()
}
}
\keyword{runShinyApps}
